pyproject.toml
[tool.poetry]
name = "odyssey"
version = "0.1.0"
description = ""
authors = ["John Halazonetis <john.halazonetis@scandit.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
psycopg = "^3.1.16"
psycopg-binary = "^3.1.16"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
